refactor: moved all components to the home folder
